Demonstrate usage of all variable types
